works on a mac

+++ src/main/scala/util/Commands.scala
@@ -13,6 +13,54 @@ case class Directory(value: String)
 case class LogEntry(value: String)
 
 class Commands {
+  val FFMPEG =
+    sys.env.get("FFMPEG") match {
+      case Some(x: String) => x
+      case None => "ffmpeg"
+    }
+
+  val NODE =
+    sys.env.get("NODE") match {
+      case Some(x: String) => x
+      case None => "node"
+    }
+
+  val YOUTUBE_DL =
+    sys.env.get("YOUTUBE_DL") match {
+      case Some(x: String) => x
+      case None => "youtube-dl"
+    }
+
+  val postmarkKey =
+    sys.env.get("POSTMARK_API_KEY") match {
+      case Some(x: String) => x
+      case None => ""
+    }
+
+  val POSTMARK_BCC =
+    sys.env.get("POSTMARK_BCC") match {
+      case Some(x: String) => x
+      case None => ""
+    }
+
+  val POSTMARK_REPLY_TO =
+    sys.env.get("POSTMARK_REPLY_TO") match {
+      case Some(x: String) => x
+      case None => ""
+    }
+
+  val CURL =
+    sys.env.get("CURL") match {
+      case Some(x: String) => x
+      case None => "curl"
+    }
+
+  val slash = java.io.File.separator
+  val quote = slash match {
+    case "/" => ""
+    case "\\" => "\""
+  }
+
   def command(cmd: String): Iterable[LogEntry] = {
     import scala.sys.process._
 
@@ -40,7 +88,7 @@ class Commands {
     import sys.process._
 
     val entireCommand =
-      "node " + cmd + " " + args.mkString(" ")
+      s"${NODE} " + cmd + " " + args.mkString(" ")
 
     println(entireCommand)
 
@@ -57,11 +105,11 @@ class Commands {
 
   def vttToSrt(dir: Directory)(id: YtId): Iterable[LogEntry] = {
 
-    val srt = dir.value + "\\v" + id.value + ".srt" // TODO use the right OS type for file strings
-    val vtt = dir.value + "\\v" + id.value + ".en.vtt"
+    val srt = dir.value + slash + "v" + id.value + ".srt"
+    val vtt = dir.value + slash + "v" + id.value + ".en.vtt"
 
-    val subtitlecmd = // TODO configurable paths
-      "\"d:/Software/ffmpeg-20160619-5f5a97d-win32-static/bin/ffmpeg.exe\" -i \"" + vtt + "\" \"" + srt + "\""
+    val subtitlecmd =
+      quote + FFMPEG + quote + " -i " + quote + vtt + quote + " " + quote + srt + quote
 
     if (!Files.exists(Paths.get(srt))) {
       command(subtitlecmd)
@@ -98,28 +146,32 @@ class Commands {
   }
 
   def youtubeDL(directory: Directory)(url: YtUrl) = {
+    println("Saving to " + directory)
+
     command(
-      "d:\\Software\\youtube-dl.exe --skip-download \"" + url.value + "\" " +
+      YOUTUBE_DL + " --skip-download " + quote + url.value + quote + " " +
         "--sub-format srt --write-sub --write-auto-sub --ignore-errors --youtube-skip-dash-manifest  " +
-        " -o \"" + directory.value + "/v%(id)s\" --write-info-json --write-description " +
+        " -o " + quote + directory.value + slash + "v%(id)s" + quote + " --write-info-json --write-description " +
         "--write-annotations --sub-lang en --no-call-home"
     )
 
-    parseJson(directory.value + "/v" + url.id.value + ".info.json")
+    parseJson(directory.value + slash + "v" + url.id.value + ".info.json")
   }
 
   def curl(directory: Directory)(url: String, filename: String): String = {
     import scala.sys.process._
 
-    val command = ("curl -o  \"" + directory.value + File.separator + filename + "\" \"" + url + "\"")
+    val command =
+      CURL + "-o  \"" + directory.value + File.separator + filename + "\" \"" + url + "\""
+
     println(command)
 
     command.!!
   }
 
   def load(dir: Directory, file: String): Iterator[String] = {
-    println(dir.value + "\\" + file)
-    Source.fromFile(dir.value + "\\" + file).getLines()
+    println(dir.value + slash + file)
+    Source.fromFile(dir.value + slash + file).getLines()
   }
 
   def email(
@@ -135,47 +187,49 @@ class Commands {
     import com.github.sebrichards.postmark.PostmarkMessage
     import com.github.sebrichards.postmark.PostmarkSuccess
 
-    val client = new PostmarkClient("895b990f-ac62-47a0-a984-a380edd59d54")
+    if (postmarkKey != "") {
+      val client = new PostmarkClient(postmarkKey)
 
-    val message = PostmarkMessage(
-      To = to,
-      From = from,
-      Subject = subject,
-      TextBody = text,
-      HtmlBody = html,
+      val message = PostmarkMessage(
+        To = to,
+        From = from,
+        Subject = subject,
+        TextBody = text,
+        HtmlBody = html,
 
-      // Optional mail fields
-      //Cc = Some("Another Recipient <another.recipient@domain.com>"),
-      Bcc = Some("gary@garysieling.com"),
-      ReplyTo = Some("gary@garysieling.com"),
+        // Optional mail fields
+        //Cc = Some("Another Recipient <another.recipient@domain.com>"),
+        Bcc = Some(POSTMARK_BCC),
+        ReplyTo = Some(POSTMARK_REPLY_TO),
 
-      // Optional attachments
-      // Attachment(new File("picture.jpg"))
-      Attachments = attachments.map(Attachment(_)), //List(
-      //Attachment("Text File.txt", "text/plain", Base64.encodeBase64String("Hello world".getBytes)),
-      //
-      //      ),
+        // Optional attachments
+        // Attachment(new File("picture.jpg"))
+        Attachments = attachments.map(Attachment(_)), //List(
+        //Attachment("Text File.txt", "text/plain", Base64.encodeBase64String("Hello world".getBytes)),
+        //
+        //      ),
 
-      // Optional Postmark fields
-      //Tag = Some("My Tag"),
-      /*Headers = List(
+        // Optional Postmark fields
+        //Tag = Some("My Tag"),
+        /*Headers = List(
         NameValueMap("key", "value"),
         NameValueMap("key2", "value2")
       ),*/
-      TrackOpens = true
-    )
+        TrackOpens = true
+      )
 
-    val result: Either[PostmarkError, PostmarkSuccess] = client.send(message)
+      val result: Either[PostmarkError, PostmarkSuccess] = client.send(message)
 
-    println(result)
+      println(result)
 
-    client.destroy
+      client.destroy
+    }
   }
 
   def save(dir: Directory, filename: String, text: String) = {
-    println(dir.value + "\\" + filename)
+    println(dir.value + slash + filename)
 
-    val file = new File(dir.value + "\\" + filename)
+    val file = new File(dir.value + slash + filename)
 
     val p = new java.io.PrintWriter(file)
     try {
@@ -189,13 +243,13 @@ class Commands {
     import java.io.{ BufferedInputStream, FileInputStream, FileOutputStream }
     import java.util.zip.{ ZipEntry, ZipOutputStream }
 
-    println(dir.value + "\\" + filename)
+    println(dir.value + slash + filename)
 
-    val zip = new ZipOutputStream(new FileOutputStream(dir.value + "\\" + filename))
+    val zip = new ZipOutputStream(new FileOutputStream(dir.value + slash + filename))
 
     include.foreach { name =>
       zip.putNextEntry(new ZipEntry(name))
-      val in = new BufferedInputStream(new FileInputStream(dir.value + "\\" + name))
+      val in = new BufferedInputStream(new FileInputStream(dir.value + slash + name))
       var b = in.read()
       while (b > -1) {
         zip.write(b)
@@ -209,6 +263,6 @@ class Commands {
   }
 
   def text(dir: Directory, file: String) = {
-    this.node("text.js", List(dir.value + "\\" + file))
+    this.node("text.js", List(dir.value + slash + file))
   }
 }
